feat: Extracted message-handler shared state, streaming handlers, and t…
- src/webview/handlers/handler-state.ts
- src/webview/handlers/streaming-handlers.ts
- src/webview/handlers/tool-execution-handlers.ts
- src/webview/message-handler.ts
- src/webview/__tests__/accessibility.test.ts

GSD-Task: S02/T01
